Bring over fixed 3.31.1; USE GCC 9.
GCC 10 changes a default setting that is a serious challenge to work around here; the best solution is to use GCC 9.x for now. I did add two RNGs to the user section, for testing; Splitmix64 is already GPL-compatible (GPL2 with Classpath Exception), and Tangle is Apache-2.0-licensed in most of the places I've put it. I'm fine with considering Tangle in the public domain as long as it isn't forced to be GPL.