Who likes deterministic test behavior? I DO!
In the released version of the original dieharder, when -S is non-zero, the seed is set, the timing tests are run, and then the actual tests start. Surprisingly, no one seemed to notice that with the same seed on the same test, different "Seed" values were shown and the results were always different between runs. The problem was in the timing tests, which seem to call the random number generator a variable amount of times, and so change the seed to some point on its cycle non-deterministically. Whoops, I guess? Pretty big whoops. This fixes that behavior by reseeding after the timing tests.